+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Prevent Oversight DAO from vetoing proposals that include a change to the Oversight DAO address [#275](https://github.com/atomone-hub/atomone/pull/275) [#299](https://github.com/atomone-hub/atomone/pull/299)
 - Set chain-wide `MaxCommissionRate` and `MinCommissionRate` to 5% and update validator commission rates accordingly [#300](https://github.com/atomone-hub/atomone/pull/300)
 - Prevent Oversight DAO change to be bundled in proposals [#316](https://github.com/atomone-hub/atomone/pull/316)
+- Migrate `x/dynamicfee` from Atom One to Atom One SDK [#322](https://github.com/atomone-hub/atomone/pull/322)
 
 ### STATE BREAKING